requirements.txt
discord.py
python-dotenv
numpy
pyaudio
openai